requirements.txt
decorator==5.1.1
Django==4.1.4
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.14.0
djangorestframework==3.14.0
firebase-admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
gunicorn==20.1.0
h5py==3.8.0
httplib2==0.21.0
importlib-metadata==5.1.0
importlib-resources==5.12.0
joblib==1.2.0
keras==2.11.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
numpy==1.24.2
pandas==1.5.3
parso==0.8.3
pefile==2023.2.7
python-dotenv==1.0.0
scikit-learn==1.2.2
seaborn==0.12.2
Send2Trash==1.8.0
speedtest==0.0.1
speedtest-cli==2.1.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.1
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.1
tensorflow-io-gcs-filesystem==0.31.0
websocket-client==1.4.2
whitenoise==6.4.0
